+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Basic Types
+  - added `HashMap.filter()`
+
 ## v0.2.4
 
 Released Sunday, 20th September 2020.